Some modifications have been made keep one step-definitions file for US05 and other for US10
